Create a classroom class and with its realtion with student class
