+++ capitalize.rb
@@ -1,0 +1,8 @@
+require './decorator'
+
+# Prepare a capitalize decorator class
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
