+++ classroom.rb
@@ -1,0 +1,13 @@
+class Classroom
+    attr_accessor :label
+    attr_reader :students
+    def initialize(label)
+        @label = label
+        @students = []
+    end
+
+    def add_student(student)
+        @students.push(student)
+        student.classroom = self
+    end
+end
