+++ decorator.rb
@@ -1,0 +1,15 @@
+require './nameable'
+
+# Prepare a base decorator class
+class BaseDecorator < Nameable
+  attr_accessor :nameable
+
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
