+++ student.rb
@@ -1,9 +1,12 @@
 require './person'
 
 class Student < Person
+  attr_accessor :classroom
+
   def initialize(age, classroom, name = 'Unknown', parent_permission: true)
     super(age, name, parent_permission)
     @classroom = classroom
+    classroom.students.push(self) unless students.include? (self)
   end
 
   def play_hooky
